fix: article delete 수정
- article 삭제시 articleArchive 먼저 삭제(추후 cascade 등 추가 예정)
- delete 시 성공 json 반환

+++ src/articles/articles.service.ts
@@ -274,7 +274,18 @@ export class ArticlesService {
    * 아티클 삭제
    */
   async remove(articleId: number): Promise<void> {
-    const article = await this.findOne(articleId);
+    const article = await this.articleRepository.findOne({ articleId }, {
+      populate: ['archives']
+    });
+
+    if (!article) {
+      throw new NotFoundException('아티클을 찾을 수 없습니다.');
+    }
+
+    if (article.archives.length > 0) {
+      await this.em.removeAndFlush(article.archives);
+    }
+
     await this.em.removeAndFlush(article);
   }
